start

+++ src/controllers/notesController.js
@@ -5,23 +5,25 @@ export const getAllNotes = async (req, res, next) => {
   try {
     const { page = 1, perPage = 10, search, tag } = req.query;
     const skip = (page - 1) * perPage;
-
+    const filter = { userId: req.user._id };
     const notesQuery = Note.find();
 
     if (search) {
-      notesQuery.where({
-        $text: { $search: search },
-      });
+      filter.$text = { $search: search };
+      // notesQuery.where({
+      //   filter.$text: { $search: search },
+      // });
     }
 
     if (tag) {
-      notesQuery.where('tag').equals(tag);
+       filter.tag = tag;
+      // notesQuery.where('tag').equals(tag);
     }
 
     // Виконуємо запити паралельно
     const [totalItems, notes] = await Promise.all([
       notesQuery.clone().countDocuments(),
-      notesQuery.skip(skip).limit(perPage),
+      notesQuery(filter).skip(skip).limit(perPage),
     ]);
 
     const totalPages = Math.ceil(totalItems / perPage);
@@ -41,12 +43,12 @@ export const getAllNotes = async (req, res, next) => {
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
 
+
   try {
-    const note = await Note.findById(noteId);
+    const note = await Note.findOne({ _id: noteId, userId: req.user._id });
 
     if (!note) {
-      next(getNotFoundById(noteId));
-      return;
+      return next(createHttpError(404, 'Note not found'));
     }
 
     res.status(200).json(note);
@@ -57,7 +59,9 @@ export const getNoteById = async (req, res, next) => {
 
 export const createNote = async (req, res, next) => {
   try {
-    const newNote = await Note.create(req.body);
+    const newNote = await Note.create({
+      ...req.body,
+    userId: req.user._id,});
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -68,11 +72,10 @@ export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
 
   try {
-    const deletedNote = await Note.findOneAndDelete({ _id: noteId });
+    const deletedNote = await Note.findOneAndDelete({ _id: noteId,  userId: req.user._id, });
 
     if (!deletedNote) {
-      next(getNotFoundById(noteId));
-      return;
+      return next(createHttpError(404, 'Note not found'));
     }
 
     res.status(200).json(deletedNote);
@@ -90,8 +93,7 @@ export const updateNote = async (req, res, next) => {
     });
 
     if (!updatedNote) {
-      next(getNotFoundById(noteId));
-      return;
+     return next(createHttpError(404, 'Note not found'));
     }
 
     res.status(200).json(updatedNote);
@@ -100,7 +102,7 @@ export const updateNote = async (req, res, next) => {
   }
 };
 
-// Допоміжна функція для 404
-const getNotFoundById = (noteId) => {
-  return createHttpError(404, `Note not found by id ${noteId}`);
-};
+// // Допоміжна функція для 404
+// const getNotFoundById = (noteId) => {
+//   return createHttpError(404, `Note not found by id ${noteId}`);
+// };
